+++ python/cudf/cudf/tests/conftest.py
@@ -276,6 +276,16 @@ def reduction_methods(request):
     return request.param
 
 
+@pytest.fixture(params=["linear", "lower", "higher", "midpoint", "nearest"])
+def quantile_interpolation(request):
+    return request.param
+
+
+@pytest.fixture(params=["spearman", "pearson"])
+def corr_method(request):
+    return request.param
+
+
 signed_integer_types = ["int8", "int16", "int32", "int64"]
 unsigned_integer_types = ["uint8", "uint16", "uint32", "uint64"]
 float_types = ["float32", "float64"]
@@ -437,6 +447,12 @@ def dropna(request):
     return request.param
 
 
+@pytest.fixture(params=[True, False])
+def skipna(request):
+    """Param for `skipna` argument"""
+    return request.param
+
+
 @pytest.fixture(params=[True, False, None])
 def nan_as_null(request):
     """Param for `nan_as_null` argument"""
@@ -458,4 +474,10 @@ def ignore_index(request):
 @pytest.fixture(params=[True, False])
 def ascending(request):
     """Param for `ascending` argument"""
+    return request.param
+
+
+@pytest.fixture(params=[True, False])
+def numeric_only(request):
+    """Param for `numeric_only` argument"""
     return request.param
