+++ python/cudf/cudf/tests/series/methods/test_unique.py
@@ -36,3 +36,51 @@ def test_datetime_unique(data, nulls):
         pd.Series(expected).sort_values(ignore_index=True),
         got.sort_values(ignore_index=True).to_pandas(),
     )
+
+
+def test_series_unique():
+    rng = np.random.default_rng(seed=0)
+    size = 100
+    arr = rng.integers(low=-1, high=10, size=size)
+    mask = arr != -1
+    sr = cudf.Series(arr)
+    sr[~mask] = None
+    assert set(arr[mask]) == set(sr.unique().dropna().to_numpy())
+    assert len(set(arr[mask])) == sr.nunique()
+
+
+def test_series_nunique(request, nan_as_null, dropna):
+    # We remove nulls as opposed to NaNs using the dropna parameter,
+    # so to test against pandas we replace NaN with another discrete value
+    request.applymarker(
+        pytest.mark.xfail(
+            nan_as_null is None,
+            reason=f"{nan_as_null=} returns wrong result",
+        )
+    )
+    cudf_series = cudf.Series([1, 2, 2, 3, 3], nan_as_null=nan_as_null)
+    pd_series = pd.Series([1, 2, 2, 3, 3])
+    expect = pd_series.nunique(dropna=dropna)
+    got = cudf_series.nunique(dropna=dropna)
+    assert expect == got
+
+    cudf_series = cudf.Series(
+        [1.0, 2.0, 3.0, np.nan, None], nan_as_null=nan_as_null
+    )
+    if nan_as_null is True:
+        pd_series = pd.Series([1.0, 2.0, 3.0, np.nan, None])
+    else:
+        pd_series = pd.Series([1.0, 2.0, 3.0, -1.0, None])
+
+    expect = pd_series.nunique(dropna=dropna)
+    got = cudf_series.nunique(dropna=dropna)
+    assert expect == got
+
+    cudf_series = cudf.Series([1.0, np.nan, np.nan], nan_as_null=nan_as_null)
+    if nan_as_null is True:
+        pd_series = pd.Series([1.0, np.nan, np.nan])
+    else:
+        pd_series = pd.Series([1.0, -1.0, -1.0])
+    expect = pd_series.nunique(dropna=dropna)
+    got = cudf_series.nunique(dropna=dropna)
+    assert expect == got
