--- python/cudf/cudf/tests/series/test_reductions.py
@@ -1,17 +1,0 @@
-# Copyright (c) 2019-2025, NVIDIA CORPORATION.
-
-import numpy as np
-import pandas as pd
-import pytest
-
-from cudf import Series
-
-
-@pytest.mark.parametrize("data", [[], [1, 2, 3]])
-def test_series_pandas_methods(data, reduction_methods):
-    arr = np.array(data)
-    sr = Series(arr)
-    psr = pd.Series(arr)
-    np.testing.assert_equal(
-        getattr(sr, reduction_methods)(), getattr(psr, reduction_methods)()
-    )
